+++ facebook/__init__.py
@@ -1,12 +1,13 @@
 #!/usr/bin/env python
 #
 # Copyright 2010 Facebook
+# Copyright 2015 Mobolic
 #
 # Licensed under the Apache License, Version 2.0 (the "License"); you may
 # not use this file except in compliance with the License. You may obtain
 # a copy of the License at
 #
-#     http://www.apache.org/licenses/LICENSE-2.0
+#     https://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
@@ -19,7 +20,7 @@
 This client library is designed to support the Graph API and the
 official Facebook JavaScript SDK, which is the canonical way to
 implement Facebook authentication. Read more about the Graph API at
-http://developers.facebook.com/docs/api.
+https://developers.facebook.com/docs/graph-api.
 
 """
 
@@ -44,14 +45,13 @@ __version__ = version.__version__
 
 FACEBOOK_GRAPH_URL = "https://graph.facebook.com/"
 FACEBOOK_OAUTH_DIALOG_URL = "https://www.facebook.com/dialog/oauth?"
-VALID_API_VERSIONS = ["2.0", "2.1", "2.2", "2.3", "2.4", "2.5", "2.6"]
+VALID_API_VERSIONS = ["2.2", "2.3", "2.4", "2.5", "2.6", "2.7", "2.8"]
 
 
 class GraphAPI(object):
     """A client for the Facebook Graph API.
 
-    See http://developers.facebook.com/docs/api for complete
-    documentation for the API.
+    https://developers.facebook.com/docs/graph-api
 
     The Graph API is made up of the objects in Facebook (e.g., people,
     pages, events, photos) and the connections between them (e.g.,
@@ -65,11 +65,11 @@ class GraphAPI(object):
        friends = graph.get_connections(user["id"], "friends")
 
     You can see a list of all of the objects and connections supported
-    by the API at http://developers.facebook.com/docs/reference/api/.
+    by the API at https://developers.facebook.com/docs/graph-api/reference/.
 
     You can obtain an access token via OAuth or by using the Facebook
     JavaScript SDK. See
-    http://developers.facebook.com/docs/authentication/ for details.
+    https://developers.facebook.com/docs/facebook-login for details.
 
     If you are using the JavaScript SDK, you can use the
     get_user_from_cookie() method below to get the OAuth access token
@@ -78,13 +78,14 @@ class GraphAPI(object):
     """
 
     def __init__(self, access_token=None, timeout=None, version=None,
-                 proxies=None):
+                 proxies=None, session=None):
         # The default version is only used if the version kwarg does not exist.
-        default_version = "2.0"
+        default_version = VALID_API_VERSIONS[0]
 
         self.access_token = access_token
         self.timeout = timeout
         self.proxies = proxies
+        self.session = session or requests.Session()
 
         if version:
             version_regex = re.compile("^\d\.\d$")
@@ -103,7 +104,7 @@ class GraphAPI(object):
 
     def get_object(self, id, **args):
         """Fetches the given object from the graph."""
-        return self.request(self.version + "/" + id, args)
+        return self.request("{0}/{1}".format(self.version, id), args)
 
     def get_objects(self, ids, **args):
         """Fetches all of the given object from the graph.
@@ -117,7 +118,7 @@ class GraphAPI(object):
     def get_connections(self, id, connection_name, **args):
         """Fetches the connections for given object."""
         return self.request(
-            "%s/%s/%s" % (self.version, id, connection_name), args)
+            "{0}/{1}/{2}".format(self.version, id, connection_name), args)
 
     def put_object(self, parent_object, connection_name, **data):
         """Writes the given object to the graph, connected to the given parent.
@@ -133,19 +134,14 @@ class GraphAPI(object):
             post = feed["data"][0]
             graph.put_object(post["id"], "comments", message="First!")
 
-        See http://developers.facebook.com/docs/api#publishing for all
-        of the supported writeable objects.
-
-        Certain write operations require extended permissions. For
-        example, publishing to a user's feed requires the
-        "publish_actions" permission. See
-        http://developers.facebook.com/docs/publishing/ for details
-        about publishing permissions.
+        Certain operations require extended permissions. See
+        https://developers.facebook.com/docs/facebook-login/permissions
+        for details about permissions.
 
         """
         assert self.access_token, "Write operations require an access token"
         return self.request(
-            self.version + "/" + parent_object + "/" + connection_name,
+            "{0}/{1}/{2}".format(self.version, parent_object, connection_name),
             post_args=data,
             method="POST")
 
@@ -159,10 +155,10 @@ class GraphAPI(object):
         being posted to the Wall. It should be a dictionary of the form:
 
             {"name": "Link name"
-             "link": "http://www.example.com/",
+             "link": "https://www.example.com/",
              "caption": "{*actor*} posted a new review",
              "description": "This is a longer description of the attachment",
-             "picture": "http://www.example.com/thumbnail.jpg"}
+             "picture": "https://www.example.com/thumbnail.jpg"}
 
         """
         return self.put_object(profile_id, "feed", message=message,
@@ -178,11 +174,11 @@ class GraphAPI(object):
 
     def delete_object(self, id):
         """Deletes the object with the given ID from the graph."""
-        self.request(self.version + "/" + id, method="DELETE")
+        self.request("{0}/{1}".format(self.version, id), method="DELETE")
 
     def delete_request(self, user_id, request_id):
         """Deletes the Request with the given ID for the given user."""
-        self.request("%s_%s" % (request_id, user_id), method="DELETE")
+        self.request("{0}_{1}".format(request_id, user_id), method="DELETE")
 
     def put_photo(self, image, album_path="me/photos", **kwargs):
         """
@@ -193,7 +189,7 @@ class GraphAPI(object):
 
         """
         return self.request(
-            self.version + "/" + album_path,
+            "{0}/{1}".format(self.version, album_path),
             post_args=kwargs,
             files={"source": image},
             method="POST")
@@ -202,7 +198,7 @@ class GraphAPI(object):
         """Fetches the current version number of the Graph API being used."""
         args = {"access_token": self.access_token}
         try:
-            response = requests.request(
+            response = self.session.request(
                 "GET",
                 FACEBOOK_GRAPH_URL + self.version + "/me",
                 params=args,
@@ -228,25 +224,29 @@ class GraphAPI(object):
         arguments.
 
         """
-        args = args or {}
 
         if post_args is not None:
             method = "POST"
 
+        # Add `access_token` to post_args or args if it has not already been
+        # included.
         if self.access_token:
-            if post_args is not None:
+            # If post_args exists, we assume that args either does not exists
+            # or it does not need `access_token`.
+            if post_args and "access_token" not in post_args:
                 post_args["access_token"] = self.access_token
-            else:
+            elif "access_token" not in args:
                 args["access_token"] = self.access_token
 
         try:
-            response = requests.request(method or "GET",
-                                        FACEBOOK_GRAPH_URL + path,
-                                        timeout=self.timeout,
-                                        params=args,
-                                        data=post_args,
-                                        proxies=self.proxies,
-                                        files=files)
+            response = self.session.request(
+                method or "GET",
+                FACEBOOK_GRAPH_URL + path,
+                timeout=self.timeout,
+                params=args,
+                data=post_args,
+                proxies=self.proxies,
+                files=files)
         except requests.HTTPError as e:
             response = json.loads(e.read())
             raise GraphAPIError(response)
@@ -274,14 +274,6 @@ class GraphAPI(object):
             raise GraphAPIError(result)
         return result
 
-    def fql(self, query):
-        """FQL query.
-
-        Example query: "SELECT affiliations FROM user WHERE uid = me()"
-
-        """
-        return self.request(self.version + "/" + "fql", {"q": query})
-
     def get_app_access_token(self, app_id, app_secret, offline=False):
         """
         Get the application's access token as a string.
@@ -291,13 +283,13 @@ class GraphAPI(object):
         access-tokens#apptokens>
         """
         if offline:
-            return "%s|%s" % (app_id, app_secret)
+            return "{0}|{1}".format(app_id, app_secret)
         else:
             args = {'grant_type': 'client_credentials',
                     'client_id': app_id,
                     'client_secret': app_secret}
 
-            return self.request("oauth/access_token",
+            return self.request("{0}/oauth/access_token".format(self.version),
                                 args=args)["access_token"]
 
     def get_access_token_from_code(
@@ -314,13 +306,14 @@ class GraphAPI(object):
             "client_id": app_id,
             "client_secret": app_secret}
 
-        return self.request("oauth/access_token", args)
+        return self.request(
+            "{0}/oauth/access_token".format(self.version), args)
 
     def extend_access_token(self, app_id, app_secret):
         """
         Extends the expiration time of a valid OAuth access token. See
-        <https://developers.facebook.com/roadmap/offline-access-removal/
-        #extend_token>
+        <https://developers.facebook.com/docs/facebook-login/access-tokens/
+        expiration-and-extension>
 
         """
         args = {
@@ -330,13 +323,14 @@ class GraphAPI(object):
             "fb_exchange_token": self.access_token,
         }
 
-        return self.request("oauth/access_token", args=args)
+        return self.request("{0}/oauth/access_token".format(self.version),
+                            args=args)
 
     def debug_access_token(self, token, app_id, app_secret):
         """
         Gets information about a user access token issued by an app. See
-        <https://developers.facebook.com/docs/facebook-login/access-tokens
-        #debug>
+        <https://developers.facebook.com/docs/facebook-login/
+        access-tokens/debugging-and-error-handling>
 
         We can generate the app access token by concatenating the app
         id and secret: <https://developers.facebook.com/docs/
@@ -345,9 +339,9 @@ class GraphAPI(object):
         """
         args = {
             "input_token": token,
-            "access_token": "%s|%s" % (app_id, app_secret)
+            "access_token": "{0}|{1}".format(app_id, app_secret)
         }
-        return self.request("/debug_token", args=args)
+        return self.request(self.version + "/" + "debug_token", args=args)
 
 
 class GraphAPIError(Exception):
@@ -391,10 +385,8 @@ def get_user_from_cookie(cookies, app_id, app_secret):
     requests to the Graph API. If the user is not logged in, we
     return None.
 
-    Download the official Facebook JavaScript SDK at
-    http://github.com/facebook/connect-js/. Read more about Facebook
-    authentication at
-    http://developers.facebook.com/docs/authentication/.
+    Read more about Facebook authentication at
+    https://developers.facebook.com/docs/facebook-login.
 
     """
     cookie = cookies.get("fbsr_" + app_id, "")
@@ -444,7 +436,7 @@ def parse_signed_request(signed_request, app_secret):
         return False
 
     # HMAC can only handle ascii (byte) strings
-    # http://bugs.python.org/issue5285
+    # https://bugs.python.org/issue5285
     app_secret = app_secret.encode('ascii')
     payload = payload.encode('ascii')